perf(parser)!: store Position line/column as u32
This keeps Location a bit more compact, which gains us a few percent points on parse
speed.

BREAKING CHANGE: `Position::line` and `Position::column` are now `u32` instead of `usize`.

+++ acdc-editor-wasm/src/lib.rs
@@ -26,8 +26,8 @@ pub struct ParseResult {
 pub struct EditorWarning {
     pub message: String,
     pub advice: Option<String>,
-    pub line: Option<usize>,
-    pub column: Option<usize>,
+    pub line: Option<u32>,
+    pub column: Option<u32>,
 }
 
 /// Initialize panic hook and set up the editor DOM orchestration.
@@ -118,7 +118,7 @@ pub fn parse_and_render(input: &str) -> Result<ParseResult, String> {
     })
 }
 
-fn location_line_col(loc: Option<&acdc_parser::SourceLocation>) -> (Option<usize>, Option<usize>) {
+fn location_line_col(loc: Option<&acdc_parser::SourceLocation>) -> (Option<u32>, Option<u32>) {
     let Some(loc) = loc else {
         return (None, None);
     };
